refactor: move out sql
